BSG-1 : Move logic to class-plugin.php.
This still isn't a class.
But it has most of the main plugin logic.
